requirements.txt
#--- Core App Dependencies ---
streamlit==1.38.0
sqlalchemy==2.0.44
python-dotenv==1.0.1
pandas==2.3.3
openpyxl==3.1.2
pyarrow==15.0.2

